# Conflicts:
#	src/main/java/com/project/doctor_fish_back/controller/DoctorController.java

+++ src/main/java/com/project/doctor_fish_back/dto/response/review/RespReviewLikeInfoDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Data
 public class RespReviewLikeInfoDto {
     private Long reviewLikeId;
-    private int likeCount;
+    private Long likeCount;
 }
